chore: slim down PR #4090 comments and tests
- Remove descriptive comments in C++ that restate what code does; keep
  the non-obvious constraint comment about specialization ordering.
- Trim 13 regression tests to 5: one per distinct concern (int64 sort,
  int64 argsort, uint64 sort, float NaN unchanged, bool specialization).
- Rename test file to use underscores to pass validate-test-names hook.
